IPC extraction and documentation
- Extract merge/split/trim to main/ipc-merge-split.js
- Extract misc handlers to main/ipc-misc.js (open-folder, check-ffmpeg, etc.)
- Extract updates to main/ipc-updates.js
- Extract logger handlers to main/ipc-logger.js
- Extract SD card and failed-operation handlers to main/ipc-sd-card.js
- Update README Project Structure
- Add ARCHITECTURE.md with IPC and module overview

Made-with: Cursor
